KOGTIO-5804: prefix package as org.kie.kogito - flexible-process-[quarkus|springboot]
